Change padding multiplier to double values

+++ src/FluentXamarinForms/FluentBase/FluentLayout.cs
@@ -43,7 +43,7 @@ namespace FluentXamarinForms.FluentBase
             return this as TFluent;
         }
 
-        public TFluent Padding (int left = 1, int top = 1, int right = 1, int bottom = 1)
+        public TFluent Padding (double left = 1, double top = 1, double right = 1, double bottom = 1)
         {
             this.BuilderActions.Add (layout => {
                 layout.Padding = FluentPadding.Padding (left, top, right, bottom);
@@ -52,7 +52,7 @@ namespace FluentXamarinForms.FluentBase
             return this as TFluent;
         }
 
-        public TFluent Padding (int horizontalSize = 1, int verticalSize = 1)
+        public TFluent Padding (double horizontalSize = 1, double verticalSize = 1)
         {
             this.BuilderActions.Add (layout => {
                 layout.Padding = FluentPadding.Padding (horizontalSize, verticalSize);
@@ -61,7 +61,7 @@ namespace FluentXamarinForms.FluentBase
             return this as TFluent;
         }
 
-        public TFluent Padding (int uniformSize = 1)
+        public TFluent Padding (double uniformSize = 1)
         {
             this.BuilderActions.Add (layout => {
                 layout.Padding = FluentPadding.Padding (uniformSize);
